Consolidate models and fixtures for reusability
