--- examples/chalice/models/__init__.py
@@ -1,8 +1,0 @@
-__all__ = ['Account', 'Biller', 'Card', 'Transaction', 'File', 'User']
-
-from .accounts import Account
-from .billers import Biller
-from .cards import Card
-from .files import File
-from .transactions import Transaction
-from .users import User
