--- examples/chalice/models/accounts.py
@@ -1,13 +1,0 @@
-from mongoengine import DateTimeField, Document, StringField
-from mongoengine_plus.models.helpers import uuid_field
-
-from agave.chalice.models import BaseModel
-
-
-class Account(BaseModel, Document):
-    id = StringField(primary_key=True, default=uuid_field('AC'))
-    name = StringField(required=True)
-    user_id = StringField(required=True)
-    platform_id = StringField(required=True)
-    created_at = DateTimeField()
-    deactivated_at = DateTimeField()
