+++ examples/chalice/resources/files.py
@@ -7,7 +7,7 @@ from mongoengine import DoesNotExist
 from agave.core.filters import generic_query
 
 from ...validators import FileQuery
-from ..models import File as FileModel
+from ...models import File as FileModel
 from .base import app
