+++ examples/fastapi/resources/files.py
@@ -6,7 +6,7 @@ from fastapi.responses import JSONResponse as Response
 from agave.core.filters import generic_query
 
 from ...validators import FileQuery, FileUploadValidator
-from ..models import File as FileModel
+from ...models import File as FileModel
 from .base import app
